% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr.data}
\alias{expr.data}
\title{Expression Data}
\format{
A data frame with 17341 rows and 127 columns
}
\usage{
data("expr.data")
}
\description{
Expression data for analysis
}
\examples{
data(expr.data)
}
\keyword{datasets}
