% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_data.R
\docType{data}
\name{GSEA_data}
\alias{GSEA_data}
\title{Example GSEA_data SummarizedExperiment Data}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{A list of matrices containing assay data.}
\item{rowData}{A DataFrame containing information about the rows.}
\item{colData}{A DataFrame containing information about the columns.}
}
}
\source{
This data was generated for test run
}
\usage{
data(GSEA_data)
}
\description{
This dataset contains example data in the SummarizedExperiment format.
It is used for demonstrating functions in this package.
}
\examples{
# Load the data
data(GSEA_data)


}
\keyword{datasets}
