% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_all.R
\docType{data}
\name{pathways.all}
\alias{pathways.all}
\title{Pathway Data}
\format{
A list file.
\describe{
Loaded by fgsea::gmtPathways
Should be obtained from MSigDB, such as msigdb.v2023.1.Hs.symbols.gmt
}
}
\source{
This data was generated for test run
}
\usage{
data(demo)
}
\description{
This dataset contains information about pathways.
}
\examples{
data(demo)
head(pathways.all)
names(pathways.all)

}
\keyword{datasets}
