% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{plot_bar}
\usage{
plot_bar(
  SE_data.fgsea,
  output_path,
  significat_type = "pval",
  topN = 10,
  strings = c("GOBP", "GOCC", "GOMF", "KEGG", "REACTOME", "WP")
)
}
\arguments{
\item{SE_data.fgsea}{Summarized Experiment Object. In the metadata of the
input, it must be a Summary Experiment Object and should contain the
following two objects "cor.df" and "fgseaRes".
"cor.df" is the result of SigCor. Add the cor.df to the metadata by running
SigCor. "fgseaRes" is the result of Sig2GSEA.
Add the fgseaRes to the metadata by running Sig2GSEA}

\item{output_path}{Character. Location of the output path.}

\item{significat_type}{Character. Filter or statistical significance.
"pval" for p-value and "qval" for Q-value}

\item{topN}{Integer. Number of significant functions in hte output bar plots.
Default is 10.}

\item{strings}{vector. Pathways header want to be plotted.}
}
\value{
Return bar plots to the outpath specified.
}
\description{
This function is explore the functions related to signature by a
surrogate of whole transcriptome data.
}
\examples{
data("demo")
output_path <- tempdir()
plot_bar(
SE_data.fgsea = SE_data,
output_path=output_path,
topN=10,
significat_type = "pval",
strings = c("KEGG"))
}
