% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat.R
\name{plot_heat}
\alias{plot_heat}
\title{plot_heat}
\usage{
plot_heat(
  SE_data.fgsea,
  output_path,
  significat_type = "pval",
  strings = c("GOBP", "GOCC", "GOMF", "KEGG", "REACTOME", "WP"),
  topN = 10,
  pathways.all,
  ranking.method = "stat",
  plot_out = TRUE
)
}
\arguments{
\item{SE_data.fgsea}{Summary Experiment Object. In the metadata of the input,
it must be a Summary Experiment Object and should contain the following two
objects "cor.df" and "fgseaRes". "cor.df" is the result of SigCor. Add the
cor.df to the metadata by running SigCor. "fgseaRes" is the result of
Sig2GSEA. Add the fgseaRes to the metadata by running Sig2GSEA}

\item{output_path}{Character. Location of the output path.}

\item{significat_type}{Character. Filter or statistical significance.
"pval" for p-value and "qval" for Q-value}

\item{strings}{vector. Pathways header want to be plotted.}

\item{topN}{Integer. Number of significant functions in hte output bar plots.
Default is 10.}

\item{pathways.all}{gmt file. loading by fgsea::gmtPathways. Should be
obtained from MSigDB: msigdb.v2023.1.Hs.symbols.gmt}

\item{ranking.method}{Character. "stat" (default) or "pval"}

\item{plot_out}{logic. Plot the heatmap to output_path (Default: FALSE).}
}
\value{
Return heatmaps to the outpath specified.
}
\description{
This is a plot component, suggesting the package performs Gene
Set Enrichment Analysis (GSEA). It uses the SigCor results (correlation
statistics) and/or p-values, along with the ontology profile, to identify
enriched pathways or gene sets associated with the input signature.
}
\examples{
data(GSEA_data)
output_path <- tempdir()
heatmap <- plot_heat(SE_data.fgsea=GSEA_data, output_path=output_path,
pathways.all=pathways.all, significat_type="pval", strings=c("KEGG"),
ranking.method="stat", plot_out=FALSE,topN=3)
}
