% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2Fun.R
\name{sig2Fun}
\alias{sig2Fun}
\title{sig2Fun}
\usage{
sig2Fun(
  SE_data,
  ranking.method = "stat",
  species = "human",
  cor.method = "spearman",
  pathways_all,
  output_path,
  topN = 10,
  Z.transform = FALSE,
  significat_type = "pval",
  strings = c("GOBP", "GOCC", "GOMF", "KEGG", "REACTOME", "WP"),
  plot_out = FALSE
)
}
\arguments{
\item{SE_data}{Summarized Experiment Object. Input contains "colData",
"assays", and "elementMetadata".
Use SummarizedExperiment::SummarizedExperiment to integrate your data.
The example code of constructing the input object:
SE_data <- SummarizedExperiment::SummarizedExperiment(
assays=list(abundance=as.matrix(assay.data)),
rowData=S4Vectors::DataFrame(rowData.data, row.names=rowData.data$sample_id),
colData=colData.data)
assay.data is the expression value where gene in column and sample in row
rowData is a vector of sample ID.
colData.data is the gene information containing three columns:
"ensg_id" (ensemble ID), "gene_symbol" (official gene symbol), and
"gene_biotype" (annotate the gene is "protein_coding" or not).
Use data(demo) to see the example.}

\item{ranking.method}{Character. "stat" (default) or "pval"}

\item{species}{Character. "human" (default) or "mouse"}

\item{cor.method}{Character. "spearman" (default), "pearson", or "kendall"}

\item{pathways_all}{list file. Loaded by fgsea::gmtPathways. Should be
obtained from MSigDB, such as msigdb.v2023.1.Hs.symbols.gmt}

\item{output_path}{Character. Location of the output path.}

\item{topN}{Integer. Number of significant functions in hte output bar plots.
Default is 10.}

\item{Z.transform}{logic. Do z-transform on GENE_MAT or not (default: FALSE).}

\item{significat_type}{Character. Filter or statistical significance.
"pval" for p-value and "qval" for Q-value}

\item{strings}{vector. Pathways header want to be plotted.
Default c("GOBP", "GOCC", "GOMF", "KEGG", "REACTOME", "WP") for
msigdb.v2023.1.Hs.symbols.gmt}

\item{plot_out}{logic. Plot the heatmap to output_path (Default: FALSE).}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function is explore the functions related to signature by a
surrogate of whole transcriptome data.
}
\examples{
data("demo")
output_path <- tempdir()
res <- sig2Fun(SE_data, ranking.method="stat", species="human",
cor.method="spearman", pathways.all, output_path,
topN=10, Z.transform=FALSE, significat_type="pval",
strings=c("KEGG"))
}
