% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIG_MAT}
\alias{SIG_MAT}
\title{Signature Matrix}
\format{
A data frame with 127 rows and 2 columns
}
\usage{
data("SIG_MAT")
}
\description{
A toy dataset representing signature values for samples.
}
\examples{
data(SIG_MAT)
}
\keyword{datasets}
