% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigCor.R
\name{SigCor}
\alias{SigCor}
\title{Sig2GSEA}
\usage{
SigCor(SE_data, output_path, cor.method = "spearman", Z.transform = FALSE)
}
\arguments{
\item{SE_data}{Summarized Experiment Object. Input contains "colData",
"assays", and "elementMetadata".
Use SummarizedExperiment::SummarizedExperiment to integrate your data.
The example code of constructing the input object:
SE_data <- SummarizedExperiment::SummarizedExperiment(
assays=list(abundance=as.matrix(assay.data)),
rowData=S4Vectors::DataFrame(rowData.data, row.names=rowData.data$sample_id),
colData=colData.data)
assay.data is the expression value where gene in column and sample in row
rowData is a vector of sample ID.
colData.data is the gene information containing three columns:
"ensg_id" (ensemble ID), "gene_symbol" (official gene symbol), and
"gene_biotype" (annotate the gene is "protein_coding" or not).
Use data(demo) to see the example.}

\item{output_path}{Character. Location of the output path.}

\item{cor.method}{Character. "spearman" (default), "pearson", or "kendall"}

\item{Z.transform}{logic. Do z-transform on GENE_MAT or not (default: FALSE).}
}
\value{
Return a SummarizedExperiment object containing analysis results.
SigCor will add "cor.df" to the metadata of the input object which is the
necessary input for Sig2GSEA.
}
\description{
This function is explore the functions related to signature by a
surrogate of whole transcriptome data.
}
\examples{
data("demo")
output_path <- tempdir()
SE_data.cor <- SigCor(
SE_data, cor.method = "spearman", output_path=output_path,
Z.transform = FALSE)
}
