% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigFun.R
\docType{package}
\name{sigFun}
\alias{SigFun}
\alias{SigFun-package}
\alias{sigFun}
\title{SigFun: Functional analysis of Gene Signature}
\description{
SigFun is an R package designed to streamline the analysis of transcriptomic
data in relation to specific gene signatures. It provides an automated
workflow for analyzing gene signatures and their functional implications in
transcriptomic datasets. The package integrates Gene Set Enrichment Analysis
(GSEA) with visualization tools to help researchers understand the
biological pathways and processes associated with their gene signatures of
interest.
}
\section{Main functions}{

\itemize{
\item \code{\link{sig2Fun}}: One-click function of running the pipeline
Analysis.
\item \code{\link{sigCor}}: Correlation analysis between signature and genes
\item \code{\link{sig2GSEA}}: GSEA analysis of signature
\item \code{\link{plot_bar}}: bar plot visualization of sig2GSEA results
\item \code{\link{plot_heat}}: heatmap visualization of sig2GSEA results
}
}

\section{Vignettes}{

See the package vignettes for detailed workflows:
\code{vignette('sigFun')}
}

\section{Installation}{

To install from Bioconductor, use:
\preformatted{
if (!requireNamespace('BiocManager', quietly=TRUE))
    install.packages('BiocManager')
BiocManager::install('sigFun')
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TMSWCChenglab/SigFun}
  \item Report bugs at \url{https://github.com/TMSWCChenglab/SigFun/issues}
}

}
\author{
\strong{Maintainer}: Wei-Chung Cheng \email{chenglabomics@gmail.com} (\href{https://orcid.org/0000-0002-1229-4857}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Chia-Hsin Liu [contributor]
  \item Pei-Chun Shen [contributor]
  \item Wen-Jen Lin [contributor]
}

}
\keyword{internal}
