% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t2g}
\alias{t2g}
\title{Transcript to Gene Mapping}
\format{
A data frame with 17341 rows and 3 columns
\describe{
\item{gs_name}{Gene set name or pathway name}
\item{ensembl_gene}{Ensembl gene ID, e.g., ENSG00000123456}
}
}
\usage{
data("t2g")
}
\description{
Another mapping object for transcript to gene relationships.
}
\examples{
data(t2g)
}
\keyword{datasets}
