% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pathways.R
\docType{data}
\name{test_pathways}
\alias{test_pathways}
\title{Example Sig2GSEA_input:test_pathways}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{A list of matrices containing assay data.}
\item{rowData}{A DataFrame containing information about the rows.}
\item{colData}{A DataFrame containing information about the columns.}
}
}
\source{
This data was generated for test run
}
\usage{
data(Sig2GSEA_input)
}
\description{
This dataset contains example data in the SummarizedExperiment format.
It is used for demonstrating functions in this package.
}
\examples{
# Load the data
data(Sig2GSEA_input)


}
\keyword{datasets}
