% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2GSEA_input.R
\docType{data}
\name{test_SE}
\alias{test_SE}
\title{Example sig2GSEA_input}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{A list of matrices containing assay data.}
\item{rowData}{A DataFrame containing information about the rows.}
\item{colData}{A DataFrame containing information about the columns.}
}
}
\source{
This data was generated for test run
}
\usage{
data(sig2GSEA_input)
}
\description{
This dataset contains example data in the SummarizedExperiment format.
It is used for demonstrating functions in this package.
}
\examples{
# Load the data
data(sig2GSEA_input)


}
\keyword{datasets}
